Move DatatypeTransformer to the datatype module so that it can be used by OntoDrivers as well.
Also move XSD vocabulary to datatype module and reverse dependency direction between jopa-api and datatype modules.
