+++ datatype/src/main/java/cz/cvut/kbss/jopa/datatype/exception/DatatypeMappingException.java
@@ -1,13 +1,11 @@
-package cz.cvut.kbss.jopa.datatype;
-
-import cz.cvut.kbss.jopa.exceptions.OWLPersistenceException;
+package cz.cvut.kbss.jopa.datatype.exception;
 
 /**
  * Indicates that the mapping has failed.
  * <p>
  * For example, if the literal value is not valid (e.g., an integer literal that cannot be parsed to Java {@link Integer}).
  */
-public class DatatypeMappingException extends OWLPersistenceException {
+public class DatatypeMappingException extends RuntimeException {
 
     public DatatypeMappingException(String message) {
         super(message);
