+++ datatype/src/main/java/cz/cvut/kbss/jopa/datatype/exception/UnsupportedTypeTransformationException.java
@@ -1,0 +1,29 @@
+/**
+ * Copyright (C) 2020 Czech Technical University in Prague
+ * <p>
+ * This program is free software: you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License as published by the Free Software
+ * Foundation, either version 3 of the License, or (at your option) any
+ * later version.
+ * <p>
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+ * details. You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+package cz.cvut.kbss.jopa.datatype.exception;
+
+/**
+ * Thrown when it is not possible to execute a type transformation.
+ */
+public class UnsupportedTypeTransformationException extends DatatypeMappingException {
+
+    public UnsupportedTypeTransformationException(String message) {
+        super(message);
+    }
+
+    public UnsupportedTypeTransformationException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
