+++ datatype/src/main/java/cz/cvut/kbss/jopa/datatype/xsd/XsdDateMapper.java
@@ -1,6 +1,6 @@
 package cz.cvut.kbss.jopa.datatype.xsd;
 
-import cz.cvut.kbss.jopa.datatype.DatatypeMappingException;
+import cz.cvut.kbss.jopa.datatype.exception.DatatypeMappingException;
 
 import javax.xml.datatype.XMLGregorianCalendar;
 import java.time.LocalDate;
