+++ datatype/src/main/java/cz/cvut/kbss/jopa/vocabulary/XSD.java
